pyproject.toml
[tool.poetry]
name = "python-310"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
jsonbasedb = "^1.2.1"
jsonbase = "^10.0.10"
gspread = "^5.4.0"
oauth2client = "^4.1.3"
requests-pkcs12 = "^1.14"
boto3 = "^1.24.67"
pandas = "^1.4.4"
django-googledrive-api = "^0.2.5"
discord-py-slash-command = "^4.2.1"
discord-py-interactions = "^4.3.1"
py-cord = "^2.1.3"
disnake = "^2.6.0"
discord = "^2.0.0"
Flask = "^2.2.2"
pycog3 = "^0.1.27"
xds-protos = "^0.0.11"
pytz = "^2022.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"